Added more fuctions , and updated some others , I also added main functions if u want to test , I think I will update one more update to correct lstmap function because it's not working

+++ libft/ft_bzero.c
@@ -6,7 +6,7 @@
 /*   By: zait-sli <zait-sli@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/04 12:17:09 by zait-sli          #+#    #+#             */
-/*   Updated: 2021/11/06 14:45:52 by zait-sli         ###   ########.fr       */
+/*   Updated: 2021/11/16 06:20:46 by zait-sli         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,10 +15,10 @@
 void	ft_bzero(void *s, size_t n)
 
 {
-	unsigned int	i;
+	size_t	i;
 
 	i = 0;
-	while (i < (unsigned int)n)
+	while (i < n)
 	{
 		((char *)s)[i] = '\0';
 		i++;
